+++ app/models/card.rb
@@ -1,7 +1,7 @@
 class Card < ApplicationRecord
-  include Assignable, Colored, Engageable, Eventable, Golden,
-    Mentions, Pinnable, Closeable, Readable, Searchable, Staged,
-    Statuses, Taggable, Watchable
+  include Assignable, Colored, Engageable, Entropy, Eventable,
+    Golden, Mentions, Pinnable, Closeable, Readable, Searchable,
+    Staged, Statuses, Taggable, Watchable
 
   belongs_to :collection, touch: true
   belongs_to :creator, class_name: "User", default: -> { Current.user }
