+++ app/views/cards/display/preview/_bubble.html.erb
@@ -1,6 +1,6 @@
-<div hidden class="bubble bubble--<%= style %>"
+<div hidden class="bubble bubble--<%= card.engagement_status %>"
      data-controller="bubble" data-action="turbo:morph-element->bubble#update"
-     data-bubble-reminder-before-value="<%= Card::AUTO_CLOSE_REMINDER_BEFORE.in_days.to_i %>"
+     data-bubble-reminder-before-value="<%= Card::ENTROPY_REMINDER_BEFORE.in_days.to_i %>"
      data-bubble-closes-at-value="<%= card.auto_close_at.iso8601 %>"
      data-bubble-entropy-action-value="<%= card_entropy_action(card) %>">
   <svg viewBox="0 0 200 100">
