+++ test/models/card/closeable_test.rb
@@ -26,30 +26,4 @@ class Card::CloseableTest < ActiveSupport::TestCase
     cards(:logo).update! last_active_at: 2.day.ago
     assert_equal (123-2).days.from_now, cards(:logo).auto_close_at
   end
-
-  test "auto close all due" do
-    cards(:logo, :shipping).each(&:reconsider)
-
-    cards(:logo).update!(last_active_at: 1.day.ago - collections(:writebook).auto_close_period)
-    cards(:shipping).update!(last_active_at: 1.day.from_now - collections(:writebook).auto_close_period)
-
-    assert_difference -> { Card.closed.count }, +1 do
-      Card.auto_close_all_due
-    end
-
-    assert cards(:logo).reload.closed?
-    assert_not cards(:shipping).reload.closed?
-  end
-
-  test "don't auto close those cards where the collection has no auto close period" do
-    cards(:logo, :shipping).each(&:reconsider)
-
-    collections(:writebook).update auto_close_period: nil
-
-    assert_no_difference -> { Card.closed.count } do
-      Card.auto_close_all_due
-    end
-
-    assert_not cards(:logo).reload.closed?
-  end
 end
